final commit

+++ lib/Model/exam_questions_model.dart
@@ -148,7 +148,7 @@ class QuestioListExamWise {
   String _questionDescription;
   int _questionId;
   String _questionImage;
-  int _questionMark;
+  double _questionMark;
   List<String> _questionOptions;
 
   QuestioListExamWise(
@@ -157,7 +157,7 @@ class QuestioListExamWise {
       String questionDescription,
       int questionId,
       String questionImage,
-      int questionMark,
+      double questionMark,
       List<String> questionOptions}) {
     this._examId = examId;
     this._questionAnswer = questionAnswer;
@@ -179,8 +179,8 @@ class QuestioListExamWise {
   set questionId(int questionId) => _questionId = questionId;
   String get questionImage => _questionImage;
   set questionImage(String questionImage) => _questionImage = questionImage;
-  int get questionMark => _questionMark;
-  set questionMark(int questionMark) => _questionMark = questionMark;
+  double get questionMark => _questionMark;
+  set questionMark(double questionMark) => _questionMark = questionMark;
   List<String> get questionOptions => _questionOptions;
   set questionOptions(List<String> questionOptions) =>
       _questionOptions = questionOptions;
